Separate "monitoring" functions from "device" functions.
Provides a 'mapper_monitor' object which can be instantiated.  The
mapper_admin is changed to work with a device, a monitor, or both,
performing whatever functions are needed for these cases.  The
database is now contained inside the monitor, so db functions are
given a context pointer, and the user application must get a db
pointer from a monitor before passing it in to a database query.

Patch also modifies qtmapper to match these changes.

git-svn-id: https://mt.music.mcgill.ca/svn/dot/branches/v2_max5/mappers/qtmapper@2374 e7448968-c8ac-449e-ac06-376e812423d6