Rename mapper_db_create/destroy_mapping and sort out the API for network control.
These functions are not db functions, but they do need the monitor
pointer in order to get an admin for sending admin messages.

Technically a device could do this too, but we assume the use case is
for monitors.  Add functions to link/unlink, connect/disconnect.

Connection properties can be specified by passing a
mapper_db_mapping_t structure, along with a bitflag indicating which
properties are valid.

git-svn-id: https://mt.music.mcgill.ca/svn/dot/branches/v2_max5/mappers/qtmapper@2554 e7448968-c8ac-449e-ac06-376e812423d6